backend/requirements.txt
djangorestframework
google-generativeai
python-dotenv
fitz
